requirements.prod.txt
# Production requirements - optimized for deployment
# Core Flask and Web Dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
gunicorn==21.2.0

# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AI/ML Dependencies (with NumPy compatibility fix)
numpy<2.0.0
tensorflow==2.15.0
keras==2.15.0
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
transformers==4.35.0
langchain==0.0.350
langchain-google-genai==1.0.0
google-generativeai==0.3.0
sentence-transformers==2.2.2

# CrewAI Dependencies
crewai==0.28.8
crewai-tools==0.1.6

# LangGraph Dependencies  
langgraph==0.0.38
langchain-core==0.1.0

# Database Dependencies
SQLAlchemy==2.0.23
psycopg2-binary==2.9.7  # For PostgreSQL support
redis==5.0.1

# Environment Management
python-dotenv==1.0.0

# Text-to-Speech
edge-tts==6.1.9
gtts==2.4.0

# PDF Generation
reportlab==4.0.4
matplotlib==3.7.2

# HTTP Requests
requests==2.31.0
httpx==0.25.1

# Async support
asyncio==3.4.3
aiofiles==23.2.1

# Data Processing
numpy==1.24.3
pandas==2.0.3

# NLP
nltk==3.8.1

# Monitoring and Logging
sentry-sdk[flask]==1.38.0

# Security
bcrypt==4.1.2

# Utilities
PyYAML==6.0.1
Pillow==10.1.0

# Vector Store
qdrant-client==1.6.9

# Development tools (optional in production)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.10.1
# flake8==6.1.0
